Dzhus/parent user fix (#179)
* Moved logic of parent creation to AuthController

* Removed endpoint Create for parent in ParentController

* Added method GetByUserId to parent service. Added method IsUserParentProfileOwner to parent controller.

* Added a check to see if the user is trying to update or get someone else's account.

* Fixed failed unit tests

* Added new unit tests to ParentControllerTest

* Added new unit tests for UserControllerTest

* Added unit tests for Parent and User Services

* Added handling if an error occurred while creating a parent.

+++ OutOfSchool/OutOfSchool.WebApi/Controllers/ParentController.cs
@@ -72,40 +72,6 @@ namespace OutOfSchool.WebApi.Controllers
         }
 
         /// <summary>
-        /// To create new Parent and add to the DB.
-        /// </summary>
-        /// <param name="userId">Id of the User.</param>
-        /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
-        [Authorize(Roles = "parent,admin")]
-        [HttpPost]
-        [Consumes(MediaTypeNames.Application.Json)]
-        [ProducesResponseType(StatusCodes.Status201Created)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<IActionResult> Create(string userId)
-        {
-            try
-            {
-                var dto = new ParentDTO
-                {
-                    Id = default,
-                    UserId = User.FindFirst("sub")?.Value,
-                };
-
-                var parent = await service.Create(dto).ConfigureAwait(false);
-
-                return CreatedAtAction(
-                     nameof(GetById),
-                     new { id = parent.Id, },
-                     parent);
-            }
-            catch (ArgumentException ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
-
-        /// <summary>
         /// To update Parent entity that already exists.
         /// </summary>
         /// <param name="shortUserDto">ShortUserDto object with new properties.</param>
@@ -114,12 +80,20 @@ namespace OutOfSchool.WebApi.Controllers
         [HttpPut]
         [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ShortUserDto))]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> Update(ShortUserDto shortUserDto)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
+            }
+
+            string userId = User.FindFirst("sub")?.Value;
+
+            if (userId != shortUserDto.Id)
+            {
+                return StatusCode(403, "Forbidden to update another user.");
             }
 
             return Ok(await service.Update(shortUserDto).ConfigureAwait(false));
@@ -134,10 +108,25 @@ namespace OutOfSchool.WebApi.Controllers
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> Delete(long id)
         {
             this.ValidateId(id, localizer);
+
+            var parent = await service.GetById(id).ConfigureAwait(false);
+
+            if (parent is null)
+            {
+                return NoContent();
+            }
+
+            var userHasRights = await IsUserParentProfileOwner(parent.Id).ConfigureAwait(false);
+
+            if (!userHasRights)
+            {
+                return StatusCode(403, "Forbidden to delete another parent account.");
+            }
 
             await service.Delete(id).ConfigureAwait(false);
 
@@ -166,6 +155,24 @@ namespace OutOfSchool.WebApi.Controllers
             {
                 return this.BadRequest(ex.Message);
             }
+        }
+
+        private async Task<bool> IsUserParentProfileOwner(long parentId)
+        {
+            // Parent can create/update/delete his/her account only if parent is the owner.
+            // Admin can manipulate data without checks.
+            if (User.IsInRole("parent"))
+            {
+                var userId = User.FindFirst("sub")?.Value;
+                var parent = await service.GetByUserId(userId).ConfigureAwait(false);
+
+                if (parentId != parent.Id)
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
     }
 }
